feat: extract inline capability handlers into endpoint modules
