+++ apps/api/src/handlers/legislation/corpus-routes.ts
@@ -1,55 +1,14 @@
-import { Result } from "better-result";
-import Elysia, { t } from "elysia";
+import Elysia from "elysia";
 
-import { readLegislationHandler } from "@/api/handlers/legislation/read-by-id";
-import {
-  searchLegislationBodySchema,
-  searchLegislationHandler,
-} from "@/api/handlers/legislation/search";
-import {
-  createSafeRootHandler,
-  type HandlerConfig,
-} from "@/api/lib/api-handlers";
+import readLegislation from "@/api/handlers/legislation/read-by-id";
+import searchLegislation from "@/api/handlers/legislation/search";
 import { authMacro, permissionMacro } from "@/api/lib/auth";
-import { tSafeId } from "@/api/lib/custom-schema";
 
 /**
  * Corpus-legislation routes (ingested statutes searchable via the
  * corpus index/pg-fts substrate). Namespaced under /legislation/corpus to
  * avoid colliding with the existing BOE proxy routes in routes.ts.
  */
-
-const searchLegislation = createSafeRootHandler(
-  {
-    permissions: { workspace: ["read"] },
-    mcp: { type: "capability", reason: "legal_corpus_admin" },
-    body: searchLegislationBodySchema,
-  } satisfies HandlerConfig,
-  async function* ({ body, scopedDb }) {
-    const response = yield* Result.await(
-      Result.tryPromise(
-        async () => await searchLegislationHandler(body, scopedDb),
-      ),
-    );
-    return Result.ok(response);
-  },
-);
-
-const readLegislation = createSafeRootHandler(
-  {
-    permissions: { workspace: ["read"] },
-    mcp: { type: "capability", reason: "legal_corpus_admin" },
-    params: t.Object({ documentId: tSafeId("legislationDocument") }),
-  } satisfies HandlerConfig,
-  async function* ({ params: { documentId }, scopedDb }) {
-    const response = yield* Result.await(
-      Result.tryPromise(
-        async () => await readLegislationHandler(documentId, scopedDb),
-      ),
-    );
-    return Result.ok(response);
-  },
-);
 
 export const legislationCorpusRoute = new Elysia({
   prefix: "/legislation/corpus",
