+++ apps/api/src/handlers/time-entries/routes.ts
@@ -1,101 +1,20 @@
-import { Result } from "better-result";
 import Elysia from "elysia";
 
 import batchDelete from "@/api/handlers/time-entries/batch-delete";
 import batchUpdate from "@/api/handlers/time-entries/batch-update";
 import createTimeEntry from "@/api/handlers/time-entries/create";
 import deleteTimeEntryById from "@/api/handlers/time-entries/delete-by-id";
-import {
-  exportCsvHandler,
-  exportCsvQuerySchema,
-} from "@/api/handlers/time-entries/export-csv";
-import {
-  exportLedesHandler,
-  exportLedesQuerySchema,
-} from "@/api/handlers/time-entries/export-ledes";
-import {
-  exportPdfHandler,
-  exportPdfQuerySchema,
-} from "@/api/handlers/time-entries/export-pdf";
+import exportCsv from "@/api/handlers/time-entries/export-csv";
+import exportLedes from "@/api/handlers/time-entries/export-ledes";
+import exportPdf from "@/api/handlers/time-entries/export-pdf";
 import readTimeEntries from "@/api/handlers/time-entries/read";
 import readTimeEntryById from "@/api/handlers/time-entries/read-by-id";
 import splitEntry from "@/api/handlers/time-entries/split";
 import timerStart from "@/api/handlers/time-entries/timer-start";
 import timerStop from "@/api/handlers/time-entries/timer-stop";
 import updateTimeEntryById from "@/api/handlers/time-entries/update-by-id";
-import { createSafeHandler } from "@/api/lib/api-handlers";
-import type { HandlerConfig } from "@/api/lib/api-handlers";
 import { permissionMacro, workspaceAccessMacro } from "@/api/lib/auth";
 import { invalidateQuery } from "@/api/lib/invalidate-query-macro";
-
-const exportCsv = createSafeHandler(
-  {
-    permissions: { workspace: ["read"] },
-    mcp: { type: "capability", reason: "billing_admin" },
-    query: exportCsvQuerySchema,
-  } satisfies HandlerConfig,
-  async function* ({ query, scopedDb, session, workspaceId }) {
-    const response = yield* Result.await(
-      Result.tryPromise(
-        async () =>
-          await exportCsvHandler({
-            workspaceId,
-            organizationId: session.activeOrganizationId,
-            query,
-            scopedDb,
-          }),
-      ),
-    );
-
-    return Result.ok(response);
-  },
-);
-
-const exportLedes = createSafeHandler(
-  {
-    permissions: { workspace: ["read"] },
-    mcp: { type: "capability", reason: "billing_admin" },
-    query: exportLedesQuerySchema,
-  } satisfies HandlerConfig,
-  async function* ({ query, scopedDb, session, workspaceId }) {
-    const response = yield* Result.await(
-      Result.tryPromise(
-        async () =>
-          await exportLedesHandler({
-            workspaceId,
-            organizationId: session.activeOrganizationId,
-            query,
-            scopedDb,
-          }),
-      ),
-    );
-
-    return Result.ok(response);
-  },
-);
-
-const exportPdf = createSafeHandler(
-  {
-    permissions: { workspace: ["read"] },
-    mcp: { type: "capability", reason: "billing_admin" },
-    query: exportPdfQuerySchema,
-  } satisfies HandlerConfig,
-  async function* ({ query, scopedDb, session, workspaceId }) {
-    const response = yield* Result.await(
-      Result.tryPromise(
-        async () =>
-          await exportPdfHandler({
-            workspaceId,
-            organizationId: session.activeOrganizationId,
-            query,
-            scopedDb,
-          }),
-      ),
-    );
-
-    return Result.ok(response);
-  },
-);
 
 export const timeEntriesRoute = new Elysia({
   prefix: "/time-entries/:workspaceId",
